World Creation and Tile-Based movement
Laid out the groundwork for creating the world based on tiles from a
sprite sheet. Two types of tiles exist: normal, and "solid" walls, which
the player can't move through.

The game now checks to see if the tile in front of the player is
non-solid before moving there. Mimic tile-by-tile movement from the
player to a satisfying level.

Added Config and Tool headers, containing constants and structs & helper
functions respectively.