World Exploration
The world now moves as the player reaches the corner. Added values to
the config header to allow for adjustment of these thresholds.

Added a filter to the render method, dictating that only entities on
screen are to be rendered. Very basic filtering, and object with
coordinates of screen but with edges that should still appear, will not
be rendered.

One bug was found where, due to various thresholds, the player will skip
a tile vertically and horizontally in the top right corner.