Simple House Generation & Safe Edges
Added an array of generated houses, in 3 different sizes. Currently not
very Object Oriented (house properties are stored in a simple struct)
but has room for improvement. Proof of concept, mostly. A lot of the
development will be in the procedural building generation, so I'll need
to start a good framework for Object Oriented generation, along with
better generation algorithms including different types of furniture,
etc.

In world movement, prevented such from revealing empty areas of the
environment. Map doesn't move if the player is at the edge (signified by
the shiny blocks at the edge).