Sharper Resolution & Radial Thresholds
Increased the default resolution to something close to 1680x1050 which
is divisible by 32x32. Makes fullscreen on my monitor look reeaal nice.

Tweaked the house generation to prevent two houses being generated in
the 0, 0 position.

Map-moving thresholds are now distanced away from the player, rather
than the edge of the screen. Makes easier to tweak, and not at the mercy
of aspect ratios being wack. Prior to this the greater dimension was
limited to the maximum threshold of the lesser dimension.

With that, the teleportation bug seems to have gone away. Still, it will
likely remain in the thresholds are too close to the edge of the screen.
Approach bug cautiously, for it is a temperamental specimen.

Now that's fixed, and all is glorious for the almighty sleep deity. Now
I may rest. Goddamn.