Prototype House
Added a prototype house, which will be developed over time to roughly
embody a kind of house that will, eventually, be procedurally generated.

Doing this has highlighted a few functional errors that will need
developing - more dynamic function calling will eventually make it
easier to build houses manually and, hence, automatically later down the
line. As it is, the code is *vey* hacky.

A few files were deleted in this commit. That's mostly because I never
realized they were still there - they've been out of use for a while.