Houses, Rooms and 16-Bit
HouseGenerator now does the same general idea, but using House objects.

House objects contain Rooms, which are constructed with a position and
size in mind. Will add a door position, passed by the House object, to
create an opening in the room. May subclass Room into different kinds to
hold different items.

Added support for changing the TILE_SIZE constant from 32 to 16 which,
in turn, changes size of the objects on screen to half the tile size.