Small Changes
A few errors found in Environment child-adding, but the outward effect
is still the same. Doing some general testing with small examples to
help with the big problems. For some reason, adding the
EnvironmentLayers to the Environment made them disappear. Their
positions seem to be correct, but something about their rendering onto
the Environment texture is causing a problem.